+++ consensus/src/main/java/org/apache/iotdb/consensus/IStateMachine.java
@@ -24,6 +24,7 @@ import org.apache.iotdb.common.rpc.thrift.TSStatus;
 import org.apache.iotdb.commons.consensus.ConsensusGroupId;
 import org.apache.iotdb.consensus.common.DataSet;
 import org.apache.iotdb.consensus.common.Peer;
+import org.apache.iotdb.consensus.common.Utils;
 import org.apache.iotdb.consensus.common.request.IConsensusRequest;
 
 import javax.annotation.concurrent.ThreadSafe;
@@ -76,18 +77,18 @@ public interface IStateMachine {
   void loadSnapshot(File latestSnapshotRootDir);
 
   /**
-   * given a snapshot dir, ask statemachine to provide all snapshot files.
+   * given a snapshot dir, ask statemachine to provide all snapshot files. By default, it will list
+   * all files recursively under latestSnapshotDir
    *
    * <p>DataRegion may take snapshot at a different disk and only store a log file containing file
    * paths. So statemachine is required to read the log file and provide the real snapshot file
    * paths.
    *
    * @param latestSnapshotRootDir dir where the latest snapshot sits
-   * @return List of real snapshot files. If the returned list is null, consensus implementations
-   *     will visit and add all files under this give latestSnapshotRootDir.
+   * @return List of real snapshot files.
    */
   default List<File> getSnapshotFiles(File latestSnapshotRootDir) {
-    return null;
+    return Utils.listAllRegularFilesRecursively(latestSnapshotRootDir);
   }
 
   /** An optional API for event notifications. */
